Fixes to session added on variable and some other stuff
