+++ i18n/sample/src/api/ZoweExplorerZosmfApi.i18n.json
@@ -1,4 +1,3 @@
 {
-    "getValidSession.loginImpossible": "Profile {0} is invalid. Please check your login details and try again.",
     "api.zosmfUSSApi.putContents": "Uploading USS file"
 }
