+++ src/api/ZoweExplorerZosmfApi.ts
@@ -10,14 +10,12 @@
 */
 
 import * as zowe from "@zowe/cli";
-import { ICommandArguments, ConnectionPropsForSessCfg, ISession, Session, IProfileLoaded, ITaskWithStatus } from "@zowe/imperative";
+import { Session, IProfileLoaded, ITaskWithStatus } from "@zowe/imperative";
 import { ZoweExplorerApi } from "./ZoweExplorerApi";
 import * as nls from "vscode-nls";
-import { DefaultProfileManager } from "../profiles/DefaultProfileManager";
-import { collectProfileDetails } from "../profiles/utils";
 import * as vscode from "vscode";
 import * as globals from "../globals";
-import { errorHandling } from "../utils";
+import { Profiles } from "../Profiles";
 
 // Set up localization
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
@@ -43,176 +41,29 @@ class ZosmfApiCommon implements ZoweExplorerApi.ICommon {
 
     public async getSession(profile?: IProfileLoaded): Promise<Session> {
         if (!this.session) {
-            this.session = await this.getValidSession((profile||this.profile), (profile||this.profile).name, false);
+            this.session = await Profiles.getInstance().getValidSession((profile||this.profile), (profile||this.profile).name, false);
         }
         return this.session;
     }
 
-    public async getStatus(validateProfile?: IProfileLoaded, profileType?: string, prompt?: boolean): Promise<{status: string,
-                                                                                                               session: Session | null}> {
+    public async getStatus(validateProfile?: IProfileLoaded, profileType?: string, prompt?: boolean): Promise<string> {
         // This API call is specific for z/OSMF profiles
         if (profileType === "zosmf") {
             try {
-                const validateSession = await this.getValidSession(validateProfile, validateProfile.name, prompt);
+                const validateSession = await Profiles.getInstance().getValidSession(validateProfile, validateProfile.name, prompt);
                 let sessionStatus;
                 if (validateSession) { sessionStatus = await zowe.CheckStatus.getZosmfInfo(validateSession); }
 
                 if (sessionStatus) {
-                    return {status: "active", session: validateSession};
+                    return "active";
                 } else {
-                    return {status: "inactive", session: undefined};
+                    return "inactive";
                 }
             } catch (err) {
                 throw new Error(err);
             }
         } else {
-            return {status: "unverified", session: undefined};
-        }
-    }
-
-    public async getValidSession(serviceProfile: IProfileLoaded,
-                                 profileName: string,
-                                 prompt?: boolean): Promise<Session | null> {
-        // Retrieve baseProfile
-        const baseProfile = DefaultProfileManager.getInstance().getDefaultProfile("base");
-
-        // If user exists in serviceProfile, use serviceProfile to login because it has precedence over baseProfile
-        // If no user & no baseProfile, use serviceProfile as default
-        // If no user & no token, use serviceProfile as default
-        if (serviceProfile.profile.user || !baseProfile || (!serviceProfile.profile.user && baseProfile && !baseProfile.profile.tokenValue)) {
-            if (prompt) {
-                // Select for prompting only fields which are not defined
-                const schemaArray = [];
-                if (!serviceProfile.profile.user && (!baseProfile || (baseProfile && !baseProfile.profile.user))) {
-                    if (!baseProfile || (baseProfile && !baseProfile.profile.tokenValue)) {
-                        schemaArray.push("user");
-                    }
-                }
-                if (!serviceProfile.profile.password && (!baseProfile || (baseProfile && !baseProfile.profile.password))) {
-                    schemaArray.push("password");
-                }
-                if (!serviceProfile.profile.host &&  (!baseProfile || (baseProfile && !baseProfile.profile.host))) {
-                    schemaArray.push("host");
-                    if (!serviceProfile.profile.port &&  (!baseProfile || (baseProfile && !baseProfile.profile.port)))
-                    { schemaArray.push("port"); }
-                }
-
-                try {
-                    const newDetails = await collectProfileDetails(schemaArray);
-                    for (const detail of schemaArray) { serviceProfile.profile[detail] = newDetails[detail]; }
-                } catch (error) {
-                    // tslint:disable:no-magic-numbers
-                    if (error.mDetails && error.mDetails.errorCode === 401) {
-                        if (globals.ISTHEIA) {
-                            vscode.window.showErrorMessage(error.message);
-                            this.getValidSession(serviceProfile, serviceProfile.name, true);
-                        } else {
-                            vscode.window.showErrorMessage(error.message, "Check Credentials").then(async (selection) => {
-                                if (selection) {
-                                    delete serviceProfile.profile.user;
-                                    delete serviceProfile.profile.password;
-                                    await this.getValidSession(serviceProfile, serviceProfile.name, true);
-                                }
-                            });
-                        }
-                    } else { throw error; }
-                }
-            }
-            const cmdArgs: ICommandArguments = {
-                $0: "zowe",
-                _: [""],
-                host: serviceProfile.profile.host ? serviceProfile.profile.host :
-                      (baseProfile ? baseProfile.profile.host : undefined),
-                port: serviceProfile.profile.port ? serviceProfile.profile.port :
-                      (baseProfile ? baseProfile.profile.port : 0),
-                basePath: serviceProfile.profile.basePath ? serviceProfile.profile.basePath :
-                          (baseProfile ? baseProfile.profile.basePath : undefined),
-                rejectUnauthorized: serviceProfile.profile.rejectUnauthorized !== null ?
-                                    serviceProfile.profile.rejectUnauthorized :
-                                    (baseProfile ? baseProfile.profile.rejectUnauthorized : true),
-                user: serviceProfile.profile.user ? serviceProfile.profile.user :
-                      (baseProfile ? baseProfile.profile.user : undefined),
-                password: serviceProfile.profile.password ? serviceProfile.profile.password :
-                          (baseProfile ? baseProfile.profile.password : undefined),
-                tokenType: "apimlAuthenticationToken",
-                tokenValue: baseProfile ? baseProfile.profile.tokenValue : undefined
-            };
-            try {
-                return zowe.ZosmfSession.createBasicZosmfSessionFromArguments(cmdArgs);
-            } catch (error) {
-                // When no password is entered, we should silence the error message for not providing it
-                // since password is optional in Zowe Explorer
-                if (error.message !== "Must have user & password OR base64 encoded credentials") {
-                    // tslint:disable:no-magic-numbers
-                    if (error.mDetails && error.mDetails.errorCode === 401) {
-                        if (globals.ISTHEIA) {
-                            vscode.window.showErrorMessage(error.message);
-                            this.getValidSession(serviceProfile, serviceProfile.name, true);
-                        } else {
-                            vscode.window.showErrorMessage(error.message, "Check Credentials").then(async (selection) => {
-                                if (selection) {
-                                    delete serviceProfile.profile.user;
-                                    delete serviceProfile.profile.password;
-                                    await this.getValidSession(serviceProfile, serviceProfile.name, true);
-                                }
-                            });
-                        }
-                    } else { throw error; }
-                }
-            }
-        } else if (baseProfile) {
-            // baseProfile exists, so APIML login is possible
-            const sessCfg = {
-                rejectUnauthorized: serviceProfile.profile.rejectUnauthorized != null ? serviceProfile.profile.rejectUnauthorized :
-                                                                                baseProfile.profile.rejectUnauthorized,
-                basePath: serviceProfile.profile.basePath,
-                hostname: serviceProfile.profile.host ? serviceProfile.profile.host : baseProfile.profile.host,
-                port: serviceProfile.profile.port ? serviceProfile.profile.port : baseProfile.profile.port,
-            };
-
-            const cmdArgs: ICommandArguments = {
-                $0: "zowe",
-                _: [""],
-                tokenType: "apimlAuthenticationToken",
-                tokenValue: baseProfile.profile.tokenValue
-            };
-
-            try {
-                let connectableSessCfg: ISession;
-                if (prompt) {
-                    connectableSessCfg = await ConnectionPropsForSessCfg.addPropsOrPrompt<ISession>(sessCfg,
-                        cmdArgs,
-                        { requestToken: false,
-                          doPrompting: prompt,
-                          getValuesBack: collectProfileDetails });
-                } else {
-                    connectableSessCfg = await ConnectionPropsForSessCfg.addPropsOrPrompt<ISession>(sessCfg,
-                                                                                                    cmdArgs,
-                                                                                                    { requestToken: false, doPrompting: false });
-                }
-
-                return new Session(connectableSessCfg);
-            } catch (error) {
-                // tslint:disable:no-magic-numbers
-                if (error.mDetails && error.mDetails.errorCode === 401) {
-                    if (globals.ISTHEIA) {
-                        vscode.window.showErrorMessage(error.message);
-                        this.getValidSession(serviceProfile, serviceProfile.name, true);
-                    } else {
-                        vscode.window.showErrorMessage(error.message, "Check Credentials").then(async (selection) => {
-                            if (selection) {
-                                delete serviceProfile.profile.user;
-                                delete serviceProfile.profile.password;
-                                await this.getValidSession(serviceProfile, serviceProfile.name, true);
-                            }
-                        });
-                    }
-                } else { throw error; }
-            }
-        } else {
-            // Neither baseProfile nor serviceProfile exists. It is impossible to login with the currently-provided information.
-            throw new Error(localize("getValidSession.loginImpossible",
-                "Profile {0} is invalid. Please check your login details and try again.", profileName));
+            return "unverified";
         }
     }
 }
